move engine update/consolidation into derive package
